+++ cognite/client/data_classes/data_modeling/instances.py
@@ -36,6 +36,7 @@ from typing_extensions import Self
 
 from cognite.client.data_classes._base import (
     CogniteObject,
+    CogniteResource,
     CogniteResourceList,
     T_CogniteResource,
     T_WriteClass,
@@ -516,15 +517,14 @@ class InstanceAggregationResult(DataModelingResource):
     @classmethod
     def _load(cls, resource: dict, cognite_client: CogniteClient | None = None) -> Self:
         """
-        Loads an instance from a json string or dictionary.
+        Loads an instance aggregation result from a json string or dictionary.
 
         Args:
             resource (dict): No description.
             cognite_client (CogniteClient | None): No description.
 
         Returns:
-            Self: An instance.
-
+            Self: An instance aggregation result.
         """
         return cls(
             aggregates=[AggregatedNumberedValue.load(agg) for agg in resource["aggregates"]],
@@ -550,6 +550,93 @@ class InstanceAggregationResult(DataModelingResource):
 
 class InstanceAggregationResultList(CogniteResourceList[InstanceAggregationResult]):
     _RESOURCE = InstanceAggregationResult
+
+
+class InspectionResults(CogniteResource):
+    def __init__(self, involved_views: list[ViewId] | None, involved_containers: list[ContainerId] | None) -> None:
+        self.involved_views = involved_views
+        self.involved_containers = involved_containers
+
+    @classmethod
+    def _load(cls, resource: dict, cognite_client: CogniteClient | None = None) -> Self:
+        involved_views = (
+            [ViewId.load(vid) for vid in resource["involvedViews"]] if "involvedViews" in resource else None
+        )
+        involved_containers = (
+            [ContainerId.load(cid) for cid in resource["involvedContainers"]]
+            if "involvedContainers" in resource
+            else None
+        )
+        return cls(
+            involved_views=involved_views,
+            involved_containers=involved_containers,
+        )
+
+    def dump(self, camel_case: bool = True) -> dict[str, Any]:
+        results = {}
+        if self.involved_views:
+            results["involvedViews" if camel_case else "involved_views"] = [
+                vid.dump(camel_case, include_type=True) for vid in self.involved_views
+            ]
+        if self.involved_containers:
+            results["involvedContainers" if camel_case else "involved_containers"] = [
+                cid.dump(camel_case, include_type=True) for cid in self.involved_containers
+            ]
+        return results
+
+
+@dataclass
+class InspectOperation(ABC): ...
+
+
+@dataclass
+class InvolvedContainers(InspectOperation): ...
+
+
+@dataclass
+class InvolvedViews(InspectOperation):
+    all_versions: bool = False
+
+
+class InstanceInspectResult(CogniteResource):
+    def __init__(
+        self,
+        space: str,
+        external_id: str,
+        instance_type: Literal["node", "edge"],
+        inspection_results: InspectionResults,
+    ) -> None:
+        self.space = space
+        self.external_id = external_id
+        self.instance_type = instance_type
+        self.inspection_results = inspection_results
+
+    @classmethod
+    def _load(cls, resource: dict, cognite_client: CogniteClient | None = None) -> Self:
+        return cls(
+            space=resource["space"],
+            external_id=resource["externalId"],
+            instance_type=resource["instanceType"],
+            inspection_results=InspectionResults.load(resource["inspectionResults"]),
+        )
+
+    def dump(self, camel_case: bool = True) -> dict[str, Any]:
+        return {
+            "space": self.space,
+            "externalId" if camel_case else "external_id": self.external_id,
+            "instanceType" if camel_case else "instance_type": self.instance_type,
+            "inspectionResults" if camel_case else "inspection_results": self.inspection_results.dump(camel_case),
+        }
+
+
+class InstanceInspectResultList(CogniteResourceList[InstanceInspectResult]):
+    _RESOURCE = InstanceInspectResult
+
+
+@dataclass
+class InstanceInspectResults:
+    nodes: InstanceInspectResultList
+    edges: InstanceInspectResultList
 
 
 class NodeApply(InstanceApply["NodeApply"]):